+++ src/ulfius.h
@@ -161,7 +161,7 @@ struct _u_response {
   char             * auth_realm;
   void             * binary_body;
   size_t             binary_body_length;
-  int             (* stream_callback) (void * stream_user_data, uint64_t offset, char * out_buf, size_t max);
+  ssize_t         (* stream_callback) (void * stream_user_data, uint64_t offset, char * out_buf, size_t max);
   void            (* stream_callback_free) (void * stream_user_data);
   uint64_t           stream_size;
   size_t             stream_block_size;
@@ -519,7 +519,7 @@ int ulfius_set_empty_body_response(struct _u_response * response, const uint sta
  */
 int ulfius_set_stream_response(struct _u_response * response, 
                                 const uint status,
-                                int (* stream_callback) (void * stream_user_data, uint64_t offset, char * out_buf, size_t max),
+                                ssize_t (* stream_callback) (void * stream_user_data, uint64_t offset, char * out_buf, size_t max),
                                 void (* stream_callback_free) (void * stream_user_data),
                                 uint64_t stream_size,
                                 size_t stream_block_size,
